+++ implementations/Direct3D12/texture.cpp
@@ -55,7 +55,7 @@ static void computeBufferImageTransferLayout(const agpu_texture_description &des
         footprint.Width = (uint32_t)std::max(compressedBlockWidth, (footprint.Width + compressedBlockWidth - 1) / compressedBlockWidth * compressedBlockWidth);
         footprint.Height = (uint32_t)std::max(compressedBlockHeight, (footprint.Height + compressedBlockHeight - 1) / compressedBlockHeight * compressedBlockHeight);
     
-        footprint.RowPitch = alignedTo(footprint.Width / compressedBlockWidth * compressedBlockSize, D3D12_TEXTURE_DATA_PITCH_ALIGNMENT);
+        footprint.RowPitch = (UINT)alignedTo(footprint.Width / compressedBlockWidth * compressedBlockSize, D3D12_TEXTURE_DATA_PITCH_ALIGNMENT);
         *transferRows = footprint.Height / compressedBlockHeight;
         *transferSlicePitch = footprint.RowPitch * (footprint.Height / compressedBlockHeight);
         *transferSize = (*transferSlicePitch) * footprint.Depth;
@@ -63,7 +63,7 @@ static void computeBufferImageTransferLayout(const agpu_texture_description &des
     else
     {
         auto uncompressedPixelSize = pixelSizeOfTextureFormat(description.format);
-        footprint.RowPitch = alignedTo(footprint.Width*uncompressedPixelSize, D3D12_TEXTURE_DATA_PITCH_ALIGNMENT);
+        footprint.RowPitch = (UINT)alignedTo(footprint.Width*uncompressedPixelSize, D3D12_TEXTURE_DATA_PITCH_ALIGNMENT);
         *transferRows = footprint.Height;
         *transferSlicePitch = footprint.RowPitch * footprint.Height;
         *transferSize = (*transferSlicePitch) * footprint.Depth;
